Configured collections refactor.
