+++ src/components/LibraryStats.tsx
@@ -1,41 +1,27 @@
 import * as React from "react";
 import { useState } from "react";
-import * as numeral from "numeral";
 import {
+  CollectionInventory,
   InventoryStatistics,
   LibraryStatistics,
   PatronStatistics,
 } from "../interfaces";
-import {
-  Bar,
-  BarChart,
-  ResponsiveContainer,
-  Tooltip,
-  TooltipProps,
-  XAxis,
-  YAxis,
-} from "recharts";
 import { Button } from "library-simplified-reusable-components";
 import InventoryReportRequestModal from "./InventoryReportRequestModal";
 import SingleStatListItem from "./SingleStatListItem";
-import { useAppAdmin, useAppFeatureFlags } from "../context/appContext";
+import {
+  useMayRequestInventoryReports,
+  useMayViewCollectionBarChart,
+} from "../businessRules/roleBasedAccess";
+import StatsCollectionsBarChart from "./StatsCollectionsBarChart";
+import StatsCollectionsList from "./StatsCollectionsList";
 
 export interface LibraryStatsProps {
   stats: LibraryStatistics;
   library?: string;
 }
 
-type OneLevelStatistics = { [key: string]: number };
-type TwoLevelStatistics = { [key: string]: OneLevelStatistics };
-type chartTooltipData = {
-  dataKey: string;
-  name?: string;
-  value: number | string;
-  color?: string;
-  perMedium?: OneLevelStatistics;
-};
-
-const inventoryKeyToLabelMap = {
+export const inventoryKeyToLabelMap = {
   titles: "Titles",
   availableTitles: "Available Titles",
   openAccessTitles: "Open Access Titles",
@@ -47,13 +33,11 @@ const inventoryKeyToLabelMap = {
   selfHostedTitles: "Self-Hosted Titles",
 };
 
+export const ALL_LIBRARIES_HEADING = "Dashboard for All Authorized Libraries";
+
 /** Displays statistics about patrons, licenses, and collections from the server,
  for a single library or all libraries the admin has access to. */
-const LibraryStats = (props: LibraryStatsProps) => {
-  const admin = useAppAdmin();
-  const { reportsOnlyForSysadmins } = useAppFeatureFlags();
-
-  const { stats, library } = props;
+const LibraryStats = ({ stats, library }: LibraryStatsProps) => {
   const {
     name: libraryName,
     key: libraryKey,
@@ -62,25 +46,16 @@ const LibraryStats = (props: LibraryStatsProps) => {
     patronStatistics: patrons,
   } = stats || {};
 
-  // A feature flag controls whether to show the inventory report form.
-  const inventoryReportRequestEnabled =
-    !reportsOnlyForSysadmins || admin.isSystemAdmin();
-
-  const chartItems = collections
-    ?.map(({ name, inventory, inventoryByMedium }) => ({
-      name,
-      ...inventory,
-      _by_medium: inventoryByMedium || {},
-    }))
-    .sort((a, b) => (a.name.toLowerCase() > b.name.toLowerCase() ? 1 : -1));
-
+  const showBarChart = useMayViewCollectionBarChart({ library });
+  const inventoryReportRequestEnabled = useMayRequestInventoryReports({
+    library,
+  });
+  const dashboardTitle = library
+    ? `${libraryName || libraryKey} Dashboard`
+    : ALL_LIBRARIES_HEADING;
   return (
     <div className="library-stats">
-      {library ? (
-        <h2>{libraryName || libraryKey} Statistics</h2>
-      ) : (
-        <h2>Statistics for All Libraries</h2>
-      )}
+      <h2>{dashboardTitle}</h2>
       <ul className="stats">
         <li className="stat-group">{renderPatronsGroup(patrons)}</li>
         <li className="stat-group">{renderCirculationsGroup(patrons)}</li>
@@ -92,7 +67,7 @@ const LibraryStats = (props: LibraryStatsProps) => {
           )}
         </li>
         <li className="stat-group stat-group-wide">
-          {renderCollectionsGroup(chartItems)}
+          {renderConfiguredCollections(collections, showBarChart)}
         </li>
       </ul>
     </div>
@@ -102,9 +77,14 @@ const LibraryStats = (props: LibraryStatsProps) => {
 const renderPatronsGroup = (patrons: PatronStatistics) => {
   return (
     <>
-      <h3>
-        <span className="stat-grouping-label">Patrons</span>
-      </h3>
+      <div>
+        <h3>
+          <span className="stat-grouping-label">Patrons</span>
+        </h3>
+        <div className={"stat-group-description"}>
+          Patrons currently registered in Palace
+        </div>
+      </div>
       <ul>
         <SingleStatListItem
           label="Total Patrons"
@@ -125,12 +105,19 @@ const renderPatronsGroup = (patrons: PatronStatistics) => {
     </>
   );
 };
+
 const renderCirculationsGroup = (patrons: PatronStatistics) => {
   return (
     <>
-      <h3>
-        <span className="stat-grouping-label">Circulation</span>
-      </h3>
+      <div>
+        <h3>
+          <span className="stat-grouping-label">Circulation</span>
+        </h3>
+        <div className={"stat-group-description"}>
+          The following circulation data displays real-time usage of the Palace
+          system.
+        </div>
+      </div>
       <ul>
         <SingleStatListItem
           label="Active Loans"
@@ -163,19 +150,29 @@ const renderInventoryGroup = (
           library={library}
         />
       )}
-      <h3>
-        <span className="stat-grouping-label">Inventory</span>
-        {inventoryReportsEnabled && library && (
-          <Button
-            callback={(() => setShowReportForm(true)) as any}
-            content="⬇︎"
-            title="Request an inventory report"
-            style={{ borderRadius: "50%", marginLeft: "10px" }}
-            className="inline small"
-            disabled={showReportForm}
-          />
-        )}
-      </h3>
+      <div>
+        <h3>
+          <span className="stat-grouping-label">Inventory</span>
+          {inventoryReportsEnabled && library && (
+            <Button
+              callback={(() => setShowReportForm(true)) as any}
+              content="⬇︎"
+              title="Request an inventory report"
+              style={{
+                borderRadius: "50%",
+                marginLeft: "10px",
+                marginBottom: "0",
+                marginTop: "-0.7rem",
+              }}
+              className="inline small"
+              disabled={showReportForm}
+            />
+          )}
+        </h3>
+        <div className={"stat-group-description"}>
+          Real-time item inventory.
+        </div>
+      </div>
       <ul>
         <SingleStatListItem
           label={inventoryKeyToLabelMap.titles}
@@ -207,178 +204,33 @@ const renderInventoryGroup = (
   );
 };
 
-const renderCollectionsGroup = (chartItems) => {
-  return chartItems.length === 0 ? (
-    <h3>No associated collections.</h3>
-  ) : (
+const renderConfiguredCollections = (
+  collections: CollectionInventory[],
+  showBarchart: boolean
+) => {
+  const content =
+    collections.length === 0 ? (
+      <span className="no-collections">No associated collections.</span>
+    ) : showBarchart ? (
+      <StatsCollectionsBarChart collections={collections} />
+    ) : (
+      <StatsCollectionsList collections={collections} />
+    );
+  return (
     <>
-      <h3>
-        <span className="stat-grouping-label">Collections</span>
-      </h3>
-      <ResponsiveContainer height={chartItems.length * 100 + 75} width="100%">
-        <BarChart
-          data={chartItems}
-          layout="vertical"
-          margin={{ top: 5, right: 30, left: 0, bottom: 50 }}
-        >
-          <YAxis
-            type="category"
-            dataKey="name"
-            interval={0}
-            angle={-45}
-            tick={{ dx: -20 }}
-            padding={{ top: 0, bottom: 0 }}
-            height={175}
-            width={125}
-          />
-          <XAxis type="number" />
-          <Tooltip content={<CustomTooltip />} />
-          <Bar
-            stackId="collections"
-            name={inventoryKeyToLabelMap.meteredLicenseTitles}
-            dataKey={"meteredLicenseTitles"}
-            barSize={50}
-            fill="#606060"
-          />
-          <Bar
-            stackId="collections"
-            name={inventoryKeyToLabelMap.unlimitedLicenseTitles}
-            dataKey={"unlimitedLicenseTitles"}
-            barSize={50}
-            fill="#404040"
-          />
-          <Bar
-            stackId="collections"
-            name={inventoryKeyToLabelMap.openAccessTitles}
-            dataKey={"openAccessTitles"}
-            barSize={50}
-            fill="#202020"
-          />
-        </BarChart>
-      </ResponsiveContainer>
+      <div>
+        <h3>
+          <span className="stat-grouping-label">Configured Collections</span>
+        </h3>
+        <div className={"stat-group-description"}>
+          The following collections are configured in your library's
+          implementation of the Palace system and are available to your users
+          through the Palace app.
+        </div>
+      </div>
+      {content}
     </>
   );
 };
-
-/* Customize the Rechart tooltip to provide additional information */
-export const CustomTooltip = ({
-  active,
-  payload,
-  label: collectionName,
-}: TooltipProps) => {
-  if (!active) {
-    return null;
-  }
-
-  // Nab inventory data from one of the chart payload objects.
-  // This corresponds to the Barcode `data` element for the current collection.
-  const chartItem = payload[0].payload;
-
-  const propertyCountsByMedium = chartItem._by_medium || {};
-  const mediumCountsByProperty: TwoLevelStatistics = Object.entries(
-    propertyCountsByMedium
-  ).reduce((acc, [key, value]) => {
-    Object.entries(value).forEach(([innerKey, innerValue]) => {
-      acc[innerKey] = acc[innerKey] || {};
-      acc[innerKey][key] = innerValue;
-    });
-    return acc;
-  }, {});
-  const aboveTheLineColor = "#030303";
-  const belowTheLineColor = "#A0A0A0";
-  const aboveTheLine: chartTooltipData[] = [
-    {
-      dataKey: "titles",
-      name: inventoryKeyToLabelMap.titles,
-      value: chartItem.titles,
-      perMedium: mediumCountsByProperty["titles"],
-    },
-    {
-      dataKey: "availableTitles",
-      name: inventoryKeyToLabelMap.availableTitles,
-      value: chartItem.availableTitles,
-      perMedium: mediumCountsByProperty["availableTitles"],
-    },
-    ...payload.filter(({ value }) => value > 0),
-  ].map(({ dataKey, name, value }) => {
-    const key = dataKey.toString();
-    const perMedium = mediumCountsByProperty[key];
-    return { dataKey: key, name, value, color: aboveTheLineColor, perMedium };
-  });
-  const aboveTheLineKeys = [
-    "name",
-    ...aboveTheLine.map(({ dataKey }) => dataKey),
-  ];
-  const belowTheLine = Object.entries(chartItem)
-    .filter(([key]) => !aboveTheLineKeys.includes(key))
-    .filter(([key]) => !key.startsWith("_"))
-    .map(([dataKey, value]) => {
-      const key = dataKey.toString();
-      const perMedium = mediumCountsByProperty[key];
-      return {
-        dataKey: key,
-        name: inventoryKeyToLabelMap[key],
-        value:
-          typeof value === "number"
-            ? value
-            : typeof value === "string"
-            ? value
-            : "",
-        color: belowTheLineColor,
-        perMedium,
-      };
-    });
-
-  // Render our custom tooltip.
-  return (
-    <div className="customTooltip">
-      <div className="customTooltipDetail">
-        <h1 className="customTooltipHeading">{collectionName}</h1>
-        {renderChartTooltipPayload(aboveTheLine)}
-        <hr style={{ margin: "0.5em 0.5em" }} />
-        {renderChartTooltipPayload(belowTheLine)}
-      </div>
-    </div>
-  );
-};
-
-const renderChartTooltipPayload = (payload: Partial<chartTooltipData>[]) => {
-  return payload.map(
-    ({ dataKey = "", name = "", value = "", color, perMedium = {} }) => (
-      <p key={dataKey} style={{ color }} className="customTooltipItem">
-        {!!name && <span>{name}:</span>}
-        <span> {formatNumber(value)}</span>
-        {perMediumBreakdown(perMedium)}
-      </p>
-    )
-  );
-};
-
-const perMediumBreakdown = (perMedium: OneLevelStatistics) => {
-  const perMediumLabels = Object.entries(perMedium)
-    .filter(([, count]) => count > 0)
-    .map(([medium, count]) => `${medium}: ${formatNumber(count)}`);
-  return (
-    !!perMediumLabels.length && (
-      <span className="customTooltipMediumBreakdown">
-        {` (${perMediumLabels.join(", ")})`}
-      </span>
-    )
-  );
-};
-
-export const formatNumber = (n: number | string | null): string => {
-  // Format numbers using US conventions.
-  // Else return non-numeric strings as-is.
-  // Otherwise, return an empty string.
-  return !isNaN(Number(n))
-    ? Intl.NumberFormat("en-US").format(Number(n))
-    : n === String(n)
-    ? n
-    : "";
-};
-
-export const humanNumber = (n: number): string =>
-  n ? numeral(n).format("0.[0]a") : "0";
 
 export default LibraryStats;
